+++ package.json
@@ -7,6 +7,7 @@
     "lucide-react": "^0.469.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^7.1.3",
     "react-scripts": "5.0.1",
     "web-vitals": "^4.2.4"
   },
